# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Digital Products Marketplace API
  version: "1.0.0"
  summary: Buyer and seller API for an India-first digital-products marketplace.
  description: |
    Three conventions run through every endpoint.

    **Money is always an object, never a bare number.** `{"minor": 118000,
    "currency": "INR", "display": "1180.00"}`. A bare number loses its currency
    somewhere between two services, and a float loses paise. The `minor` field
    is the authoritative integer count of minor units; `display` is for humans
    and must never be parsed back into an amount.

    **Errors are RFC 9457 problem documents.** `application/problem+json`, with a
    stable `type` URI per failure class so a client can branch without parsing
    prose.

    **Identifiers are prefixed.** `prd_`, `ord_`, `lic_`, `ast_`, `var_` followed
    by 26 Crockford base32 characters. The prefix makes a misrouted identifier
    obvious in a log; Crockford omits I, L, O and U so reading one aloud does
    not create a support case.

    This document covers the public surface. Operator endpoints — `/api/v1/admin/*`
    behind a role, and `/internal/verify/*` behind an internal token — are
    deliberately not published here: they exist, they are audited, and telling
    the world where they are serves no caller. `TestOpenAPIMatchesRoutes`
    compares this file against the router on every run, so a public endpoint
...
    fails the build.
  contact:
    name: Security contact
...
  license:
...
servers:
...
security:
  - sessionCookie: []
paths:
...
  /api/v1/catalog/products:
    get:
      tags: [Catalogue]
      summary: List products
      security: []
      parameters:
        - { name: category, in: query, schema: { type: string } }
        - { name: sort, in: query, schema: { type: string, enum: [ranked, newest, price_asc, price_desc] } }
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 100, default: 24 } }
        - { name: offset, in: query, schema: { type: integer, minimum: 0, default: 0 } }
      responses:
        "200":
          description: |
            A page of products. `ranking_disclosure_url` and `ranking_note` are
            on every listing response rather than buried on a policy page.
          content:
            application/json:
              schema:
                type: object
                required: [products, ranking_disclosure_url, ranking_note]
                properties:
                  products:
                    type: array
                    items: { $ref: "#/components/schemas/ProductSummary" }
                  sort: { type: string }
                  limit: { type: integer }
                  offset: { type: integer }
                  ranking_disclosure_url: { type: string, const: /legal/ranking }
                  ranking_note:
                    type: string
                    example: >-
                      Order is produced by a published formula. There is no paid
                      placement: no position on this marketplace can be bought.
        "429": { $ref: "#/components/responses/RateLimited" }
  /api/v1/catalog/products/{id}:
    get:
      tags: [Catalogue]
      summary: Product detail
      security: []
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, pattern: "^prd_[0-9A-HJKMNP-TV-Z]{26}$" } }
      responses:
        "200":
          description: The product, its detail and its variants
          content:
            application/json:
              schema:
                type: object
                properties:
                  product: { $ref: "#/components/schemas/ProductSummary" }
                  detail: { type: object }
                  variants:
                    type: array
                    items: { $ref: "#/components/schemas/Variant" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/catalog/search:
    get:
      tags: [Catalogue]
      summary: Search the catalogue
      security: []
      parameters:
        - { name: q, in: query, required: true, schema: { type: string, minLength: 1, maxLength: 200 } }
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 100, default: 24 } }
        - { name: offset, in: query, schema: { type: integer, minimum: 0, default: 0 } }
      responses:
        "200":
          description: Search results
          content:
            application/json:
              schema:
                type: object
                properties:
                  query: { type: string }
                  results:
                    type: array
                    items: { $ref: "#/components/schemas/ProductSummary" }
                  limit: { type: integer }
                  offset: { type: integer }
                  ranking_disclosure_url: { type: string }
        "429": { $ref: "#/components/responses/RateLimited" }

  # ----------------------------------------------------------- authentication
  /api/v1/auth/register:
    post:
      tags: [Authentication]
      summary: Create an account
      description: |
        The response is identical whether or not the address is already
        registered. Only the e-mail sent differs, because a response that differs
        is an account-enumeration oracle.

        `accept_notice_version` records consent against a specific privacy-notice
        version, which is what makes the consent record meaningful later.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password, accept_notice_version]
              properties:
                email: { type: string, format: email, maxLength: 254 }
                password:
                  type: string
                  minLength: 12
                  description: |
                    Checked against a breach list and against leet-speak and
                    digit-stripped normalisations, so `p4ssw0rd1234` is refused.
                display_name: { type: string, maxLength: 120 }
                country: { type: string, minLength: 2, maxLength: 2, example: IN }
                state_code: { type: integer, description: GST state code. }
                accept_notice_version: { type: string }
                notice_language: { type: string, example: en }
      responses:
        "201":
          description: Accepted. Verification e-mail queued.
          content:
            application/json:
              schema:
                type: object
                properties:
                  status: { type: string }
        "422": { $ref: "#/components/responses/ValidationFailed" }
        "429": { $ref: "#/components/responses/RateLimited" }
  /api/v1/auth/login:
    post:
      tags: [Authentication]
      summary: Sign in
      description: |
        A failure is identical for a missing account and a wrong password,
        including in timing: the missing-account path runs a real Argon2id hash
        so the two take the same time.

        Sellers must supply `totp_code` or `recovery_code`; production
        configuration refuses to disable that requirement, because a seller
        controls a payout destination.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string }
                totp_code: { type: string, pattern: "^[0-9]{6}$" }
                recovery_code: { type: string }
      responses:
        "200":
          description: |
            Signed in. Sets a `__Host-`-prefixed session cookie and a CSRF cookie.
          headers:
            Set-Cookie:
              schema: { type: string }
              description: "__Host-session and __Host-csrf; Secure; HttpOnly; SameSite=Lax"
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Session" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "429": { $ref: "#/components/responses/RateLimited" }
...
  /api/v1/auth/session:
    get:
      tags: [Authentication]
      summary: Current session
      description: |
        Roles are resolved fresh from the database on every request, never read
        from a client-held token, so a revocation takes effect immediately.
      responses:
        "200":
          description: The session
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Session" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /api/v1/auth/password-reset:
    post:
      tags: [Authentication]
      summary: Begin a password reset
      description: Always 202, whether or not the address exists.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email: { type: string, format: email }
      responses:
        "202": { description: If the address exists, a reset e-mail was queued. }
        "429": { $ref: "#/components/responses/RateLimited" }
  /api/v1/auth/password-reset/complete:
    post:
      tags: [Authentication]
      summary: Complete a password reset
      description: Single-use token. Completing it revokes every existing session.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token, password]
              properties:
                token: { type: string }
                password: { type: string, minLength: 12 }
      responses:
        "200": { description: Password changed; all sessions revoked. }
        "422": { $ref: "#/components/responses/ValidationFailed" }

  # ----------------------------------------------------------------- commerce
  /api/v1/checkout:
    post:
      tags: [Commerce]
      summary: Create an order
      description: |
        Prices and tax are computed **server-side**. The client sends product and
        variant identifiers only — never an amount — so there is nothing to
        tamper with.

        Each line's `explanation` narrates how every figure was derived, in plain
        language, and travels with the order so a buyer or seller never has to
        ask.
      parameters:
        - name: Idempotency-Key
          in: header
          schema: { type: string, minLength: 8, maxLength: 255, pattern: "^[A-Za-z0-9._:-]+$" }
          description: Retrying with the same key returns the same order.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [items]
              properties:
                items:
                  type: array
                  minItems: 1
                  items:
                    type: object
                    required: [product_id, variant_id]
                    properties:
                      product_id: { type: string, pattern: "^prd_[0-9A-HJKMNP-TV-Z]{26}$" }
                      variant_id: { type: string, pattern: "^var_[0-9A-HJKMNP-TV-Z]{26}$" }
                state_code: { type: integer, description: GST state code; defaults to the account's. }
                country: { type: string, example: IN }
                gstin: { type: string, description: Required when is_b2b is true. }
                is_b2b: { type: boolean, default: false }
      responses:
        "201":
          description: Order created, awaiting payment
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Order" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "422": { $ref: "#/components/responses/ValidationFailed" }
        "429": { $ref: "#/components/responses/RateLimited" }
  /api/v1/orders/{id}/payment:
    post:
      tags: [Commerce]
      summary: Begin payment
      description: |
        Opens a payment at the provider, including the split instruction where
        the provider supports it, and returns the parameters the client hands to
        the provider's checkout. The buyer pays the provider directly — money
        never passes through this platform.
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, pattern: "^ord_[0-9A-HJKMNP-TV-Z]{26}$" } }
      responses:
        "200":
          description: Provider handoff
          content:
            application/json:
              schema:
                type: object
                properties:
                  order_id: { type: string }
                  provider: { type: string, enum: [bridge, razorpay_route, mor] }
                  intent_id: { type: string }
                  amount: { $ref: "#/components/schemas/Money" }
                  checkout: { type: object, description: Provider-specific checkout parameters. }
                  expires_at: { type: string, format: date-time }
        "404": { $ref: "#/components/responses/NotFound" }
        "409":
          description: The order is not in a payable state.
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
  /api/v1/orders/{id}/confirm:
    post:
      tags: [Commerce]
      summary: Confirm a payment
      description: |
        Called when the buyer returns from the provider. The signature is
        recomputed as `HMAC-SHA256(key_secret, order_id + "|" + payment_id)`; a
        payment that does not verify is never honoured, whatever the client says.

        This races the provider's webhook by design. Both call the same capture
        routine and de-duplicate on the same idempotency key, so whichever
        arrives first does the work and the other is a no-op.
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [intent_id, payment_id, signature]
              properties:
                intent_id: { type: string }
                payment_id: { type: string }
                signature: { type: string, description: Hex HMAC-SHA256. }
      responses:
        "200":
          description: Paid. The entitlement exists and the ledger is posted.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Order" }
        "400":
          description: Signature verification failed.
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/orders:
    get:
      tags: [Commerce]
      summary: List the caller's orders
      parameters:
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 100, default: 24 } }
        - { name: offset, in: query, schema: { type: integer, minimum: 0, default: 0 } }
      responses:
        "200":
          description: Orders
          content:
            application/json:
              schema:
                type: object
                properties:
                  orders:
                    type: array
                    items: { $ref: "#/components/schemas/Order" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /api/v1/orders/{id}:
    get:
      tags: [Commerce]
      summary: Get one order
      description: Ownership is part of the query, not a check after the fetch.
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: The order
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Order" }
        "404": { $ref: "#/components/responses/NotFound" }
...
  /api/v1/library:
    get:
      tags: [Delivery]
      summary: The caller's entitlements
      responses:
        "200":
          description: Licences and their assets
          content:
            application/json:
              schema:
                type: object
                properties:
                  library:
                    type: array
                    items: { $ref: "#/components/schemas/Entitlement" }
        "401": { $ref: "#/components/responses/Unauthorized" }
...
  /api/v1/seller/kyc/submit:
    post:
      tags: [Seller]
      summary: Submit for verification
      responses:
        "202": { description: Submitted for review }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409":
          description: Already under review, or a prerequisite is missing.
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: __Host-session
      description: |
        Set at login. State-changing requests additionally require the
        `__Host-csrf` cookie echoed in the `X-CSRF-Token` header — a signed
        double-submit bound to the session, checked alongside `Sec-Fetch-Site`,
        `Origin` and `Referer`.
    metricsToken:
      type: http
      scheme: bearer
      description: "`METRICS_TOKEN`. Required in production."
    webhookSignature:
      type: apiKey
      in: header
      name: X-Razorpay-Signature
      description: HMAC-SHA256 of the raw body under the provider webhook secret.
...